Move loginusers function to account module
